trunks: ftdi: fix reset pulse generation
When testing working with the C232HM USB->FTDI->SPI probe it was
observed, that attempts to set to LOW the GPIOL0 lead (connected to
the TPM2 reset pin on the breakout board) causes the CS signal to also
go low. It turns out the driver is using the wrong template for
manipulating the GPIO pin. The 'stop' template should be used, as the
GPIO levels are not supposed to be altered while a SPI access is in
progress.

BUG=chromium:498861
TEST=verified that reset can be asserted without affecting CS.

Change-Id: I2144b1b3f6e9d25980b2e811fdbbeade05172e23
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/277164
Reviewed-by: Utkarsh Sanghi <usanghi@chromium.org>