settingsd: Document removal
This CL adds removal of SettingsDocuments in the SettingsMap interface
and implements it in SimpleSettingsMap. One necessary prerequisite for
this is to replace the unique_ptr's in SimpleSettingsMap by weak_ptr's
to keep track of the custom deleter of the shared_ptr's. Additionally it
introduces a Key class that represents all identifiers.

BUG=chromium:503587
TEST=Compiles. Unittests pass.

Change-Id: I04b275d1219db3a39ff003cc870997062974ab09
Reviewed-on: https://chromium-review.googlesource.com/280517
Commit-Queue: Christoph Schütte <cschuet@chromium.org>
Trybot-Ready: Christoph Schütte <cschuet@chromium.org>
Tested-by: Christoph Schütte <cschuet@chromium.org>
Reviewed-by: Christoph Schütte <cschuet@chromium.org>