quipper: Make GetSize() const; Redo header generation
GetSize() should be const but it relies on RegenerateHeader(), which
updates |out_header_|, the header that will be used to write perf data.

It makes no sense for GetSize() to indirectly update a member variable.
Instead, change RegenerateHeader() to GenerateHeader(), which modifies
the header passed in as an argument. This allows both functions to be
const.

BUG=chromium:476625
TEST=unit tests pass

Change-Id: Iec7c0ed4446be849a5d35636e6bc181c43e109d8
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/281784
Reviewed-by: David Sharp <dhsharp@chromium.org>