permission_broker: Make OpenPath a "normal" method to return errors.
Returning an invalid dbus::FileDescriptor is not the correct way to
signal failure in a DBus method implementation as the DBus adapters will
attempt to pass the file descriptor and it will fail because it is
invalid. This change makes OpenPath a "normal" method so that it can
respond with an error message.

BUG=chromium:496469
TEST=Triggered a open failure to validate success and failure paths.

Change-Id: Icfacdc37e78ffb822b489dc0a7a1d41a30aee794
Reviewed-on: https://chromium-review.googlesource.com/282070
Tested-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Kees Cook <keescook@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Trybot-Ready: Reilly Grant <reillyg@chromium.org>