Revert "update_engine: Change the default chunk_size to 200MB"
The chunk_size handling in payload_generator has some problems. This commit break the full payload generation due to some bugs when the chunk_size is not a divisor of the partition size, breaking paygen in canaries.

This patch is not required for chromium:504447 since the bzip2 bug was already fixed.

This reverts commit def5fe30d2c5f38915f4e5986bdc54661f20ec3e.

BUG=chromium:504447
TEST=cros_generate_update_payload for a full payload doesn't generate a 200MiB kernel image.

Change-Id: I505afd84ca10c7256b66cd84f288c6db03a4d88e
Reviewed-on: https://chromium-review.googlesource.com/282767
Reviewed-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Trybot-Ready: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>