debugd: CWP: Adjust period of TLB miss events on x86_64 CPUs
At the current sampling period, TLB miss events are not frequent enough for
us to get usable samples. Reduce the sampling period to be in approximate
ratio to the number of cycles events, based on running perf stat on some
test machines. This should result in approximately the same number of
samples in a TLB miss report as a cycles report.

We observed a cycles/[id]TLB-missed ratio of ~700 for Celeron 2955U
(on boards Peppy and Wolf).

TEST=none
BUG=none

Change-Id: I025de4c81f52d6a5c0e9e20c79e5b20c055ddd45
Reviewed-on: https://chromium-review.googlesource.com/282757
Reviewed-by: Emma Rapati <rapati@chromium.org>
Commit-Queue: David Sharp <dhsharp@chromium.org>
Tested-by: David Sharp <dhsharp@chromium.org>