power: Improve force-screen-on-for-user-activity behavior.
When the user manually sets the panel backlight brightness
to 0 using the brightness-down key and then uses the
keyboard or touchpad, powerd automatically increases the
brightness to the minimum-visible level. It was only
updating the level for the current power source, though,
meaning that if the user then switched from AC to battery
power or vice versa, the backlight would be turned off
immediately.

Make powerd instead update both the AC and battery
brightness levels to the minimum-visible level in response
to user activity.

BUG=chromium:507944
TEST=added a test; also manually verified that the screen
     stays on after setting to 0% while on AC, hitting a
     key, and then switching to battery

Change-Id: Ia46ae5844511d638d0bfdefbf722c9839a4c810a
Reviewed-on: https://chromium-review.googlesource.com/284049
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Commit-Queue: Dan Erat <derat@chromium.org>