permission_broker: Look up udev device by device node path only once.
This change optimizes the permission broker by observing that all Rules
are actually UdevRules (except for DenyClaimedUsbDeviceRule, which has
been trivially converted). This means that instead of each taking a
device path and mapping it to a udev_device object this can be done once
by the rule engine before handling any rules. This results in a MASSIVE
improvement in the amount of time it takes to process an access request.

BUG=None
TEST=Ran permission_broker unittests and verified the performance
improvement by opening a USB device from a Chrome App.

Change-Id: I3dd7087ad074897b8f8c3f3bc791df50aa751bcc
Reviewed-on: https://chromium-review.googlesource.com/284410
Tested-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Trybot-Ready: Reilly Grant <reillyg@chromium.org>