settingsd: Move SettingsDocument ownership out of SettingsMap
The idea is that SettingsMap is merely responsible for figuring out
effective settings by constructing the proper order of the relevant
SettingsDocuments and resolving overlaps. SettingsDocument ownership
will in turn require additional tasks, such as deleting documents from
disk etc. so the to-be-introduced SettingsDocumentManager class is a
better owner for our SettingsDocuments.

BUG=chromium:504784
TEST=unit tests

Change-Id: I5fb187f9e14c4e4da2de119b37a0f3054a49cf5e
Reviewed-on: https://chromium-review.googlesource.com/284800
Reviewed-by: Christoph Schütte <cschuet@chromium.org>
Commit-Queue: Mattias Nissler <mnissler@chromium.org>
Tested-by: Mattias Nissler <mnissler@chromium.org>