update_engine: Remove usage of g_shell_parse_argv().
g_shell_parse_argv() was used as a helper function to specify a command
line in unittests as a string and convert it to a list of command
arguments. With C++11, the overhead of declaring them as a vector of
arguments is minimal, so we just remove the usage of the function
g_shell_parse_argv() completely.

BUG=chromium:499886
TEST=uniitest still pass.

Change-Id: Ibb63ea2cc85d6976bfd45c496906bfcce1f421d6
Reviewed-on: https://chromium-review.googlesource.com/284890
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Alex Vakulenko <avakulenko@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Trybot-Ready: Alex Deymo <deymo@chromium.org>