settingsd: Reporting unreferenced documents
This CL modifies the SettingsMap interface by adding an additional out
parameter to mutating methods, i.e. InsertDocument() and
RemoveDocument(), to report the list of documents that have become
unreferenced as a result of the mutation.

Also it fixes an invalid deletion when adding an empty document.

BUG=chromium:503587
TEST=Compiles. Unittests pass.

Change-Id: I99414206d6adcebf9d972fceeef54220a78f4bbe
Reviewed-on: https://chromium-review.googlesource.com/285471
Reviewed-by: Christoph Schütte <cschuet@chromium.org>
Tested-by: Christoph Schütte <cschuet@chromium.org>
Commit-Queue: Christoph Schütte <cschuet@chromium.org>