shill: call WiFiObjectTest::InstallMockWakeOnWiFi in WiFiObjectTest::SetUp
Move the call to WiFiObjectTest::InstallMockWakeOnWiFi from the constructor of
WiFiObjectTest to WiFiObjectTest::SetUp. This ensures that the destructor of
the SimpleAlarmTimer in WakeOnWiFi is only invoked after the EventDispatcher
(in WiFiMainTest) has been initialized (failing to ensure this results in a
deadlock).

Please refer to the bug for more details on how calling
WiFiObjectTest::InstallMockWakeOnWiFi in the constructor of WiFiObjectTest
results in a deadlock.

BUG=chromium:509138
TEST=The following command succeeds without hanging:

  USE="asan clang" FEATURES=test P2_TEST_FILTER=shill::WiFiMainTest.* \
    emerge-$BOARD shill

Change-Id: I986ed285207480ae0a2179e98d5b881ca29bbcfb
Reviewed-on: https://chromium-review.googlesource.com/286223
Reviewed-by: Zeping Qiu <zqiu@chromium.org>
Reviewed-by: Samuel Tan <samueltan@chromium.org>
Commit-Queue: Samuel Tan <samueltan@chromium.org>
Tested-by: Samuel Tan <samueltan@chromium.org>