shill: NetlinkPacket: Add new object for netlink messages
Create a new object type for retrieving a netlink message from
a byte buffer, and consuming data from the payload.  This will
be the primary method for passing around netlink packet data.
It will also be used to replace netlink parsing in the future.

BUG=chromium:512152
TEST=New unit test

Change-Id: I21d91162af69a7e80e519d52b5fdfc1c00aee9ed
Reviewed-on: https://chromium-review.googlesource.com/287198
Reviewed-by: Samuel Tan <samueltan@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>