quipper: Use perf_option_parser in PerfRecorder to validate command lines
Quipper now requires that "perf" is the second argument (after duration),
followed by either "record" or "stat" and any of the whitelisted options
in perf_option_parser. PerfRecorder will replace "perf" with
"/usr/bin/perf" by default.

In PerfRecorderTest, "/usr/bin/perf" is replaced with
{"sudo", "/usr/bin/perf"} instead, as root is required to run perf. This
continues to work fine for running the test in the CrOS SDK.

The goal of this change is to make it safer to allow perf options that
may have come from a compromised chrome, so that quipper, when called via
debugd, does not become a route to privilege escalation.

TEST=unit tests
BUG=none

Change-Id: I68a8eafd7041b85cb344577e74d1fe13e26f32ba
Reviewed-on: https://chromium-review.googlesource.com/287432
Reviewed-by: Kees Cook <keescook@chromium.org>
Commit-Queue: David Sharp <dhsharp@chromium.org>
Tested-by: David Sharp <dhsharp@chromium.org>