power_manager: Write DBus xml interface.
The org.chromium.PowerManager interface implemented by powerd is
manually defined in the code and not autogenerated from and xml spec.
Consumers of this API defined its own xml file that may not be in sync
with the interface maintained in power_manager.

This patch moves the .xml definition of power_manager to the standard
locations in the dbus_bindings directory and updates the callers to
include it. This will be reused by other daemons in the future and
eventually powerd can use the autogenerated dbus adaptor instead of
manually dealing with the dbus interaction.

BUG=chromium:419827
TEST=`FEATURES=test USE=gdmwimax emerge-link wimax_manager` still works.

Change-Id: Iaab110d70486357a162e31f8e036b79f808f9e34
Reviewed-on: https://chromium-review.googlesource.com/287930
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Alex Vakulenko <avakulenko@chromium.org>
Reviewed-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>