shill: add MockIcmpSession, IcmpSessionFactory, and MockIcmpSessionFactory
Add MockIcmpSession class that allows for mocking of
IcmpSession function calls. Add an IcmpSessionFactory and
a corresponding Mock of this factory to allow IcmpSessions
that are dynamically allocated within a class implementation
to be mocked in unit tests.

A forthcoming CL will use these classes for testing.

BUG=None
TEST=Compiling shill for unit tests succeeds.

Change-Id: Ic327a19229d25146abf62f2228782b1826b186e7
Reviewed-on: https://chromium-review.googlesource.com/289954
Tested-by: Samuel Tan <samueltan@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Samuel Tan <samueltan@chromium.org>