power: Add support for suspend_to_idle
This adds support for suspending to idle (freeze) for systems. The powerd
pref file suspend_to_idle controls whether the system suspends to idle
(value of 1) or to memory (value of 0).

BUG=chrome-os-partner:43269
TEST=powerd_dbus_suspend on glados
check that we suspended to freeze instead of mem in /var/log/messages

Change-Id: Ib7ffe195806b0f77eb6843f5abe868b5dc843736
Reviewed-on: https://chromium-review.googlesource.com/288753
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>