shill: Connection: Fix gateway reachability using link routes
Previously, in order to fix reachability to a default gateway
(in order to apply a gateway default route), shill used one
of two methods:

 - Open up the prefix assigned to the local address until
   both the host and gateway were in the same broadcast
   domain.
 - Add the gateway as a peer address for the interface.

Although this works well in most situations, there are cases
where it doesn't work as well.  Specifically, modifying the
prefix may prevent the local host from reaching other peers
that are really not on the same broadcast domain.

Instead of modifying the prefix, this new change adds a
link-scoped route to the gateway.  The host acts identically
(ARPs for the gateway IP address locally), but since the
prefix is preserved, it sends traffic to all other addresses
outside the provided prefix through the gateway.

BUG=chromium:515156
TEST=Unit tests + manual -- trigger this code by forcing /32 prefix

Change-Id: I71e33b6ed0dfe74e069376ab8bd5710f7d30b4bd
Reviewed-on: https://chromium-review.googlesource.com/289599
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: Andrey Ulanov <andreyu@google.com>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>