power: wait briefly before executing base.reboot
If the device reboots before buffet has a change to indicate to the
cloud that the command has been completed, the device will get stuck in
a reboot loop until the command expires.

This is a work-around since a more correct solution isn't feasible in the short
term.

BUG=chrome-os-partner:43342
TEST=manually executed base.reboot command

Change-Id: Iafc162729b4742e1ead7365a3f1fe4e693c01b05
Reviewed-on: https://chromium-review.googlesource.com/290046
Reviewed-by: Dan Erat <derat@chromium.org>
Commit-Queue: Aaron Kemp <kemp@google.com>
Tested-by: Aaron Kemp <kemp@google.com>