settingsd: Keep source ID and blob handle in SettingsDocument.
This creates private members in SettingsDocument storing the source ID
and the blob handle for a given document. This is for the benefit of
SettingsDocumentManager, which needs this meta data for maintaining
documents and is now declared as a friend of SettingsDocument. While
at it, remove GetSourceID() from the virtual interface of
SettingsDocument as it doesn't belong there.

BUG=chromium:504784
TEST=unit tests pass

Change-Id: Id6f2b0bd7e592892647882cd29c6d12ebab6c027
Reviewed-on: https://chromium-review.googlesource.com/291320
Commit-Ready: Mattias Nissler <mnissler@chromium.org>
Tested-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Christoph Schütte <cschuet@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>